feat: dataStore에 저장된 값에 따라 가이드 화면 visible 로직 구현
