+++ android/app/src/main/java/com/mulberry/ody/presentation/feature/room/etadashboard/EtaDashboardFragment.kt
@@ -8,20 +8,12 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.lifecycleScope
-import com.mulberry.ody.data.local.db.OdyDataStore
 import com.mulberry.ody.presentation.feature.room.MeetingRoomViewModel
 import com.mulberry.ody.presentation.theme.OdyTheme
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.launch
-import javax.inject.Inject
 
 @AndroidEntryPoint
 class EtaDashboardFragment : Fragment() {
-    @Inject
-    lateinit var odyDatastore: OdyDataStore
-
     private val viewModel: MeetingRoomViewModel by activityViewModels<MeetingRoomViewModel>()
 
     override fun onCreateView(
@@ -42,32 +34,6 @@ class EtaDashboardFragment : Fragment() {
                 }
             }
         }
-    }
-
-    override fun onViewCreated(
-        view: View,
-        savedInstanceState: Bundle?,
-    ) {
-        super.onViewCreated(view, savedInstanceState)
-        initializeGuide()
-    }
-
-    private fun initializeGuide() {
-        lifecycleScope.launch {
-            val isFirstSeenEtaDashboard = odyDatastore.getIsFirstSeenEtaDashboard().first()
-            if (isFirstSeenEtaDashboard) {
-                startEtaDashboardGuide()
-                odyDatastore.setIsFirstSeenEtaDashboard(false)
-            }
-        }
-    }
-
-    private fun startEtaDashboardGuide() {
-//        childFragmentManager.commit {
-//            add(R.id.fcv_eta_dashboard, EtaDashboardGuideFirstFragment())
-//            setReorderingAllowed(true)
-//            addToBackStack(null)
-//        }
     }
 
     private fun onBack() {
